partselect_ai_backend/pyproject.toml
[project]
name = "partselect-ai-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastapi>=0.115.12",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-deepseek>=0.1.3",
    "langchain-google-community>=2.0.7",
    "langgraph>=0.3.27",
    "openai>=1.72.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]
